chore(e2e): make browsers run in parallel when executing tests
related to OPT-2898

+++ client/e2e/setup.js
@@ -9,12 +9,7 @@ import fetch from 'node-fetch';
 import config from './config';
 import license from './license';
 
-export default async function setup() {
-  await ensureLicense();
-  await cleanEntities();
-}
-
-async function ensureLicense() {
+export async function ensureLicense() {
   const resp = await fetch(config.endpoint + '/api/license/validate');
   if (!resp.ok) {
     await fetch(config.endpoint + '/api/license/validate-and-store', {
@@ -24,25 +19,34 @@ async function ensureLicense() {
   }
 }
 
-async function getSession() {
+async function getSession(user) {
   const resp = await fetch(config.endpoint + '/api/authentication', {
     method: 'POST',
     headers: {'Content-Type': 'application/json'},
-    body: JSON.stringify(config.agentUser)
+    body: JSON.stringify(user)
   });
   return await resp.text();
 }
 
-async function cleanEntities() {
-  const headers = {Cookie: `X-Optimize-Authorization="Bearer ${await getSession()}"`};
+export async function cleanEntities(ctx) {
+  if (ctx.users) {
+    for (let i = 0; i < ctx.users.length; i++) {
+      const headers = {
+        Cookie: `X-Optimize-Authorization="Bearer ${await getSession(ctx.users[i])}"`
+      };
 
-  const response = await fetch(`${config.endpoint}/api/entities`, {headers});
-  const entities = await response.json();
+      const response = await fetch(`${config.endpoint}/api/entities`, {headers});
+      const entities = await response.json();
 
-  for (let i = 0; i < entities.length; i++) {
-    await fetch(`${config.endpoint}/api/${entities[i].entityType}/${entities[i].id}?force=true`, {
-      method: 'DELETE',
-      headers
-    });
+      for (let i = 0; i < entities.length; i++) {
+        await fetch(
+          `${config.endpoint}/api/${entities[i].entityType}/${entities[i].id}?force=true`,
+          {
+            method: 'DELETE',
+            headers
+          }
+        );
+      }
+    }
   }
 }
